+++ tests/integration/update.js
@@ -199,6 +199,54 @@ module.exports = (session) => {
           .catch(done);
       });
 
+      it('should use `Model.createValidationError` to create the error', done => {
+        class MyError extends Error {
+          constructor(errors) {
+            super('MyError');
+            this.errors = errors;
+          }
+        }
+
+        let ModelWithSchema = subClassWithSchema(Model1, {
+          type: 'object',
+          properties: {
+            id: {type: ['number', 'null']},
+            model1Prop1: {type: 'string'},
+            model1Prop2: {type: 'number'}
+          }
+        });
+
+        ModelWithSchema.createValidationError = (errors) => {
+          return new MyError(errors);
+        };
+
+        ModelWithSchema
+          .query()
+          .update({model1Prop1: 666})
+          .then(() => {
+            done(new Error('should not get here'));
+          })
+          .catch(err => {
+            expect(err).to.be.a(MyError);
+            expect(err.errors).to.eql({
+              model1Prop1: [{ 
+                message: 'should be string',
+                keyword: 'type',
+                params: {
+                  type: 'string'
+                }
+              }]
+            });
+            
+            return session.knex(Model1.tableName);
+          })
+          .then(rows => {
+            expect(_.map(rows, 'model1Prop1').sort()).to.eql(['hello 1', 'hello 2', 'hello 3']);
+            done();
+          })
+          .catch(done);
+      });
+
     });
 
     describe('.query().updateAndFetchById()', () => {